# openapi.yaml
openapi: "3.0.3"
info:
  title: mcp-scan API
  version: "0.1.0"
  description: |
    REST API for the mcp-scan MCP server security scanner.

    Use this API to:
    - Trigger security scans of local MCP server projects
    - Poll for scan status and retrieve structured results
    - Configure webhooks to receive real-time notifications on scan events

    All endpoints return JSON. Errors follow the `ApiError` schema with an
    HTTP-idiomatic `statusCode`, a machine-readable `error` code, and a
    human-readable `message`.

    **Pagination** — list endpoints accept `page` (1-based) and `pageSize`
    (1–100, default 20) query parameters. Response envelopes include `total`,
    `page`, and `pageSize`. Pagination metadata is also echoed in response
    headers (`X-Total-Count`, `X-Page`, `X-Page-Size`).

    **Rate limiting** — every response includes `X-RateLimit-*` headers.
    Production deployments should enforce limits at the gateway layer.
  contact:
    name: Syntrophy Foundry
    url: https://syntrophy.ai
  license:
    name: MIT

servers:
  - url: http://localhost:3001
    description: Local development server

tags:
  - name: scans
    description: Scan lifecycle — initiate, track, and retrieve results
  - name: webhooks
    description: Webhook registration for async event notifications
  - name: health
    description: Server health and readiness

paths:
  # ---------------------------------------------------------------------------
  # Health
  # ---------------------------------------------------------------------------
  /health:
    get:
      operationId: getHealth
      tags: [health]
      summary: Health check
      description: Returns server health status. Use for liveness and readiness probes.
      responses:
        "200":
          description: Server is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              example:
                status: ok
                version: "0.1.0"
                timestamp: "2026-03-29T20:00:00.000Z"

  # ---------------------------------------------------------------------------
  # Scans
  # ---------------------------------------------------------------------------
  /scans:
    post:
      operationId: createScan
      tags: [scans]
      summary: Initiate a scan
      description: |
        Enqueue a new scan job. The scan runs asynchronously; the response
        returns immediately with status `queued` and a job `id`.

        Poll `GET /scans/{id}` to track progress, or supply a `webhookUrl` to
        receive a `scan.completed` / `scan.failed` notification when the scan
        finishes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScanRequest"
            examples:
              basic:
                summary: Scan a local directory
                value:
                  target: /path/to/my-mcp-server
              with_rules:
                summary: Scan with specific rules only
                value:
                  target: /path/to/my-mcp-server
                  rules: ["MCP-001", "MCP-002"]
              with_webhook:
                summary: Scan with webhook notification
                value:
                  target: /path/to/my-mcp-server
                  webhookUrl: https://example.com/hooks/mcp-scan
      responses:
        "202":
          description: Scan job accepted and queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanJob"
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                target: "/path/to/my-mcp-server"
                status: queued
                createdAt: "2026-03-29T20:00:00.000Z"
        "400":
          $ref: "#/components/responses/BadRequest"

    get:
      operationId: listScans
      tags: [scans]
      summary: List scan history
      description: Returns a paginated list of all scan jobs, newest first.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: status
          in: query
          description: Filter by job status
          required: false
          schema:
            $ref: "#/components/schemas/ScanJobStatus"
      responses:
        "200":
          description: Paginated list of scan jobs
          headers:
            X-Total-Count:
              description: Total number of matching jobs
              schema:
                type: integer
            X-Page:
              description: Current page number
              schema:
                type: integer
            X-Page-Size:
              description: Items per page
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedScanJobs"
        "400":
          $ref: "#/components/responses/BadRequest"

  /scans/{scanId}:
    parameters:
      - $ref: "#/components/parameters/ScanId"
    get:
      operationId: getScan
      tags: [scans]
      summary: Get scan status
      description: Returns the current state of a scan job including metadata.
      responses:
        "200":
          description: Scan job record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanJob"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteScan
      tags: [scans]
      summary: Delete a scan record
      description: |
        Remove a completed or failed scan from history.
        Returns 409 if the scan is still running.
      responses:
        "204":
          description: Scan record deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

  /scans/{scanId}/results:
    parameters:
      - $ref: "#/components/parameters/ScanId"
    get:
      operationId: getScanResults
      tags: [scans]
      summary: Retrieve scan results
      description: |
        Returns the full `ScanResult` payload once the scan has completed.

        Returns `202` with a `SCAN_IN_PROGRESS` error body if the scan is still
        running. Returns `422` with a `SCAN_FAILED` error body if the scan
        encountered a fatal error.
      responses:
        "200":
          description: Scan completed — full results returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResult"
        "202":
          description: Scan still in progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanInProgressError"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: Scan failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ---------------------------------------------------------------------------
  # Webhooks
  # ---------------------------------------------------------------------------
  /webhooks:
    post:
      operationId: createWebhook
      tags: [webhooks]
      summary: Register a webhook
      description: |
        Register an HTTPS endpoint to receive event notifications.

        Supported events:
        - `scan.completed` — fired when a scan finishes successfully
        - `scan.failed` — fired when a scan encounters a fatal error
        - `scan.critical_finding` — fired when a completed scan has ≥1 `error`-severity finding

        If `secret` is provided, each delivery includes an
        `X-MCP-Scan-Signature: sha256=<hmac>` header for verification.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWebhookRequest"
            examples:
              all_events:
                summary: Subscribe to all events
                value:
                  url: https://example.com/hooks/mcp-scan
                  events: ["scan.completed", "scan.failed", "scan.critical_finding"]
                  secret: my-webhook-secret
              critical_only:
                summary: Critical findings only
                value:
                  url: https://example.com/hooks/mcp-scan-critical
                  events: ["scan.critical_finding"]
      responses:
        "201":
          description: Webhook registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookConfig"
        "400":
          $ref: "#/components/responses/BadRequest"

    get:
      operationId: listWebhooks
      tags: [webhooks]
      summary: List registered webhooks
      responses:
        "200":
          description: Array of registered webhooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WebhookConfig"

  /webhooks/{webhookId}:
    parameters:
      - $ref: "#/components/parameters/WebhookId"
    get:
      operationId: getWebhook
      tags: [webhooks]
      summary: Get a webhook
      responses:
        "200":
          description: Webhook config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookConfig"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteWebhook
      tags: [webhooks]
      summary: Unregister a webhook
      responses:
        "204":
          description: Webhook removed
        "404":
          $ref: "#/components/responses/NotFound"

# =============================================================================
# Components
# =============================================================================
components:
  parameters:
    ScanId:
      name: scanId
      in: path
      required: true
      description: UUID of the scan job
      schema:
        type: string
        format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"

    WebhookId:
      name: webhookId
      in: path
      required: true
      description: UUID of the webhook registration
      schema:
        type: string
        format: uuid

    Page:
      name: page
      in: query
      required: false
      description: 1-based page number
      schema:
        type: integer
        minimum: 1
        default: 1

    PageSize:
      name: pageSize
      in: query
      required: false
      description: Items per page (max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"

    Conflict:
      description: Operation conflicts with current state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"

  schemas:
    # -------------------------------------------------------------------------
    # Core domain types
    # -------------------------------------------------------------------------
    Severity:
      type: string
      enum: [error, warning, note, none]
      description: Severity level of a security finding

    Finding:
      type: object
      required: [ruleId, message, severity, filePath]
      properties:
        ruleId:
          type: string
          example: MCP-001
        message:
          type: string
          example: "Prompt Injection Vector: Tool description contains suspicious instruction override pattern."
        severity:
          $ref: "#/components/schemas/Severity"
        filePath:
          type: string
          example: /path/to/my-mcp-server/src/tools.ts
        line:
          type: integer
          minimum: 1
          example: 42
        column:
          type: integer
          minimum: 1
          example: 7
        snippet:
          type: string
          example: 'description: "ignore previous instructions and..."'

    FindingSummary:
      type: object
      required: [errors, warnings, notes]
      properties:
        errors:
          type: integer
          minimum: 0
        warnings:
          type: integer
          minimum: 0
        notes:
          type: integer
          minimum: 0

    ScanResult:
      type: object
      required: [target, startedAt, finishedAt, findings, filesScanned, errors, summary]
      properties:
        target:
          type: string
          description: Absolute path that was scanned
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        findings:
          type: array
          items:
            $ref: "#/components/schemas/Finding"
        filesScanned:
          type: integer
          minimum: 0
        errors:
          type: array
          description: Non-fatal errors encountered during scanning
          items:
            type: string
        summary:
          $ref: "#/components/schemas/FindingSummary"

    ScanJobStatus:
      type: string
      enum: [queued, running, completed, failed]

    ScanJob:
      type: object
      required: [id, target, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        target:
          type: string
          description: Path or URL that was scanned
        rules:
          type: array
          items:
            type: string
          description: Rule IDs applied, or null for all rules
          nullable: true
        status:
          $ref: "#/components/schemas/ScanJobStatus"
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        result:
          $ref: "#/components/schemas/ScanResult"
          nullable: true
        error:
          type: string
          nullable: true
          description: Error message when status is 'failed'

    PaginatedScanJobs:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ScanJob"
        total:
          type: integer
          minimum: 0
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1

    # -------------------------------------------------------------------------
    # Request bodies
    # -------------------------------------------------------------------------
    CreateScanRequest:
      type: object
      required: [target]
      properties:
        target:
          type: string
          description: Local directory path or URL of the MCP server to scan
          example: /path/to/my-mcp-server
        rules:
          type: array
          items:
            type: string
          description: Optional list of rule IDs to apply. Omit to run all rules.
          example: ["MCP-001", "MCP-003"]
        webhookUrl:
          type: string
          format: uri
          description: Optional one-off URL to notify on scan completion or failure
          example: https://example.com/hooks/mcp-scan

    WebhookEvent:
      type: string
      enum: [scan.completed, scan.failed, scan.critical_finding]
      description: |
        - `scan.completed` — scan finished successfully
        - `scan.failed` — scan encountered a fatal error
        - `scan.critical_finding` — scan completed with at least one error-severity finding

    WebhookConfig:
      type: object
      required: [id, url, events, createdAt]
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            $ref: "#/components/schemas/WebhookEvent"
        secret:
          type: string
          description: HMAC secret — included in response for reference; treat as sensitive
          nullable: true
        createdAt:
          type: string
          format: date-time

    CreateWebhookRequest:
      type: object
      required: [url, events]
      properties:
        url:
          type: string
          format: uri
          description: HTTPS endpoint to deliver events to
        events:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/WebhookEvent"
        secret:
          type: string
          description: Optional shared secret for HMAC signature verification

    # -------------------------------------------------------------------------
    # Error schemas
    # -------------------------------------------------------------------------
    ApiError:
      type: object
      required: [error, message, statusCode]
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: NOT_FOUND
        message:
          type: string
          description: Human-readable description
          example: "Scan job '550e8400...' not found."
        statusCode:
          type: integer
          example: 404

    ScanInProgressError:
      type: object
      required: [error, message, statusCode, scanId, status]
      properties:
        error:
          type: string
          example: SCAN_IN_PROGRESS
        message:
          type: string
          example: "Scan '550e8400...' is still running. Poll GET /scans/{id} for status."
        statusCode:
          type: integer
          example: 202
        scanId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/ScanJobStatus"

    HealthResponse:
      type: object
      required: [status, version, timestamp]
      properties:
        status:
          type: string
          enum: [ok]
        version:
          type: string
        timestamp:
          type: string
          format: date-time